edit action on applied jobs

+++ bde/views.py
@@ -168,6 +168,7 @@ def remove_lead_generated(request):
         print(request.POST['check'])
         update_contract_tbl = Contract.objects.get(id=request.POST['check'])
         update_contract_tbl.lead_generated = '0'
+        update_contract_tbl.invited_by_client = '0'
         update_contract_tbl.save()
         remove_lead = LeadGeneratedContract.objects.get(lead_contract_id=update_contract_tbl.id)
         remove_lead.delete()
@@ -237,7 +238,85 @@ def add_sales_info(request):
 
 
 def delete_job(request):
-    print(request.POST['project_id'])
-    contract = Contract.objects.get(id=request.POST['project_id'])
-    contract.delete()
-    return render(request, "bde/applied_jobs.html")
+    if request.session.has_key('username'):
+        print(request.POST['project_id'])
+        contract = Contract.objects.get(id=request.POST['project_id'])
+        contract.delete()
+        return render(request, "bde/applied_jobs.html")
+    else:
+        return redirect("/")
+
+
+def edit_contract_info(request):
+    if request.session.has_key('username'):
+        print(request.GET['id'])
+        contract_info = Contract.objects.get(id=request.GET['id'])
+        project_size = ProjectSize.objects.all()
+        freelance_account = FreelanceAccount.objects.all()
+        return render(request, "bde/edit_contract_info.html",
+                      {'contract_info': contract_info, 'project_sizes': project_size,
+                       'freelance_accounts': freelance_account})
+    else:
+        return redirect("/")
+
+
+def edit_contract_info_form_submission(request):
+    if request.session.has_key('username'):
+        selected_contract = Contract.objects.get(id=request.POST['contract_id'])
+        if 'attachment' in request.FILES:
+            print("in if")
+            attachment = request.FILES['attachment']
+            fs = FileSystemStorage()
+            ext = os.path.splitext(attachment.name)[1]
+            ext = ext.lower()
+            print(ext)
+            date = datetime.now()
+            result = '%s%s%s%s%s%s_%s' % (
+                date.year, date.month, date.day, date.hour, date.minute, date.second, os.urandom(10).hex())
+            print(result)
+            filename = fs.save(result + ext, attachment)
+            upload_file_url = fs.url(filename)
+            selected_contract.project_url = upload_file_url
+
+        selected_contract.project_title = request.POST['project_title']
+        selected_contract.project_url = request.POST['project_url']
+        selected_contract.project_size_id = request.POST['project_size']
+        selected_contract.freelance_account_id = request.POST['freelance_account']
+        selected_contract.client_location = request.POST['client_location']
+        selected_contract.project_desc = request.POST['project_description']
+        selected_contract.contract_type = request.POST['payment_method']
+        selected_contract.contract_price = request.POST['project_amount']
+        bde_user = User.objects.get(user_email=request.session['username'])
+
+        if selected_contract.invited_by_client == '1' and request.POST['invited_by_client'] == '0':
+            lead_generated_contract = LeadGeneratedContract.objects.get(lead_contract_id=selected_contract.id)
+            lead_generated_contract.delete()
+            selected_contract.lead_generated = '0'
+            selected_contract.invited_by_client = request.POST['invited_by_client']
+            # selected_contract.save()
+        elif selected_contract.invited_by_client == '0' and request.POST['invited_by_client'] == '1':
+            lead_generated_contract = LeadGeneratedContract(lead_contract_id=selected_contract.id,
+                                                            bidder_account_id=bde_user.id)
+            lead_generated_contract.save()
+            selected_contract.lead_generated = '1'
+            selected_contract.invited_by_client = request.POST['invited_by_client']
+            # selected_contract.save()
+
+        selected_contract.save()
+
+        filter_string = request.session['bde_selectedFreelancer']
+        if filter_string == 'all':
+            bde_user = User.objects.get(user_email=request.session['username'])
+            jobs_list = Contract.objects.filter(bidder_account_id=bde_user.id).order_by('-created_at')
+            freelance_account = FreelanceAccount.objects.all();
+        else:
+            bde_user = User.objects.get(user_email=request.session['username'])
+
+            jobs_list = Contract.objects.filter(bidder_account_id=bde_user.id,
+                                                freelance_account_id=filter_string).order_by('-created_at')
+            freelance_account = FreelanceAccount.objects.all();
+        return redirect("/bde/applied_jobs",
+                      {'jobs_lists': jobs_list, 'freelance_accounts': freelance_account,
+                       'selected_option': filter_string})
+    else:
+        return redirect("/")
